--- components/Modal/modal.d.ts
@@ -1,8 +1,0 @@
-import { ReactNode } from "react";
-
-export type Modal = {
-  children: ReactNode | any;
-  title: string;
-  open: boolean;
-  onClose: () => void;
-};
