+++ components/page-actions.tsx
@@ -1,44 +1,27 @@
 import { GetStaticProps } from "next";
 import React, { useEffect, useState } from "react";
+import { Modal } from "react-bootstrap";
 import IssueAvatars from "./issue-avatars";
-import Modal from "./Modal";
+import NewProposal from "./new-proposal";
 
 export default function PageActions() {
-  const [open, setOpen] = useState(false);
-
-  function handleOpenClick() {
-    setOpen(true);
-  }
-  function handleCloseClick() {
-    setOpen(false);
-  }
-
   return (
-    <>
-      <div className="container">
-        <div className="row justify-content-center">
-          <div className="col-md-10">
-            <div className="d-flex align-items-center justify-content-between mb-4">
-              <h4 className="h4">Details</h4>
-              <div className="d-flex align-items-center">
-                <IssueAvatars></IssueAvatars>
-                <button className="btn btn-md btn-opac mr-1">
-                  View on github
-                </button>
-                <button
-                  className="btn btn-md btn-primary"
-                  onClick={handleOpenClick}>
-                  Start working
-                </button>
-              </div>
+    <div className="container">
+      <div className="row justify-content-center">
+        <div className="col-md-10">
+          <div className="d-flex align-items-center justify-content-between mb-4">
+            <h4 className="h4">Details</h4>
+            <div className="d-flex align-items-center">
+              <IssueAvatars></IssueAvatars>
+              <button className="btn btn-md btn-opac mr-1">
+                View on github
+              </button>
+              <NewProposal />
             </div>
           </div>
         </div>
       </div>
-      <Modal title="modal-test" open={open} onClose={handleCloseClick}>
-        test
-      </Modal>
-    </>
+    </div>
   );
 }
