+++ package.json
@@ -22,7 +22,6 @@
     "next": "^11.0.0",
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.0",
-    "react-cool-portal": "^1.1.1",
     "react-dom": "^17.0.2",
     "remark": "^12.0.0",
     "remark-html": "^12.0.0",
