--- styles/components/modal.scss
@@ -1,10 +1,0 @@
-.modal {
-  &-backdrop {
-    position: fixed;
-    inset: 0;
-    background-color: opacify(black, 0.2);
-  }
-  &-dialog {
-    background-color: $dark-gray;
-  }
-}
